chore: moved components to dashboard-components folder
