+++ apps/web/app/(authenticated)/(app)/app/[apiId]/page.tsx
@@ -1,4 +1,4 @@
-import { ColumnChart } from "@/components/charts";
+import { ColumnChart } from "@/components/dashboard-components/charts";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
 import { getTenantId } from "@/lib/auth";
 import { fillRange } from "@/lib/utils";
