+++ apps/web/app/(authenticated)/auth/sign-up/layout.tsx
@@ -1,4 +1,4 @@
-import { Particles } from "@/components/particles";
+import { Particles } from "@/components/dashboard-components/particles";
 import { auth } from "@clerk/nextjs";
 import { redirect } from "next/navigation";
