+++ apps/web/components/dashboard-components/api-key-table/table.tsx
@@ -41,7 +41,7 @@ import {
   DialogTrigger,
 } from "@/components/ui/dialog";
 import { trpc } from "@/lib/trpc/client";
-import { toast } from "../ui/use-toast";
+import { toast } from "../../ui/use-toast";
 import { useRouter } from "next/navigation";
 interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[];
